chore: refactor get_pytorch_version
